Migrate into monorepo
Separate the setup wizard and the actual mangaupdates server. This
will be used for pkg when packaging the setup wizard into an executable
file.

- This monorepo solely uses yarn workspaces to separate packages.
- All type errors have been fixed and builds have been changed.
- Test files have been removed.

References #3
