+++ package.json
@@ -1,31 +1,18 @@
 {
-  "name": "manga-updates",
+  "name": "mangaupdates",
   "version": "1.0.0",
-  "main": "dist/src/index.js",
   "license": "Apache-2.0",
-  "scripts": {
-    "build": "tsc",
-    "start": "node dist/src/index.js",
-    "scrapeBackup": "node dist/src/PBBackupEncoder/index.js",
-    "setup": "node dist/src/Setup/index.js",
-    "lint": "eslint --ext .ts .",
-    "stest": "node dist/src/SourceUpdates/testSourceCode.js",
-    "mconvert": "node dist/src/ManualEncoder/index.js"
-  },
+  "private": true,
+  "workspaces": [
+    "server",
+    "setup-wizard"
+  ],
   "devDependencies": {
     "@types/node": "^15.6.1",
-    "@types/prompt-sync": "^4.1.0",
     "@typescript-eslint/eslint-plugin": "^4.25.0",
     "@typescript-eslint/parser": "^4.25.0",
     "eslint": "^7.27.0",
     "prettier": "^2.3.0",
     "typescript": "^4.3.2"
-  },
-  "dependencies": {
-    "axios": "^0.21.1",
-    "cheerio": "^1.0.0-rc.9",
-    "node-cache": "^5.1.2",
-    "prompt-sync": "^4.2.0",
-    "xmlhttprequest": "^1.8.0"
   }
 }
