+++ server/src/NotificationServices/index.ts
@@ -1,0 +1,2 @@
+export { configureIfttt } from './ifttt';
+export { configureSpontit } from './spontit';
