+++ server/src/SourceUpdates/MangaDex.ts
@@ -1,5 +1,5 @@
 import axios, { AxiosError } from 'axios';
-import { MangaPacket } from 'types/sourceEntries';
+import { MangaPacket } from '../types';
 
 export default async function fetchMangaDex(): Promise<Array<MangaPacket>> {
     const MangaDexUpdates: Array<MangaPacket> = [];
