+++ server/src/SourceUpdates/MangaKakalot.ts
@@ -1,11 +1,11 @@
 import axios, { AxiosError } from 'axios';
 import cheerio from 'cheerio';
-import { MangaPacket } from 'types/sourceEntries';
+import { MangaPacket } from '../types';
 
-export default async function fetchMangaNelo(): Promise<Array<MangaPacket>> {
+export default async function fetchMangaKakalot(): Promise<Array<MangaPacket>> {
     const mangaNeloUpdates: Array<MangaPacket> = [];
 
-    const baseDomain = 'https://manganato.com/';
+    const baseDomain = 'https://mangakakalot.com/';
     const html = await axios.get(baseDomain, { timeout: 30000 }).catch((err: AxiosError) => {
         switch (err.code) {
             case 'ECONNABORTED':
@@ -28,13 +28,12 @@ export default async function fetchMangaNelo(): Promise<Array<MangaPacket>> {
         decodeEntities: true
     });
 
-    for (const manga of $('div.content-homepage-item', 'div.panel-content-homepage').toArray()) {
-        const title = $('img', manga).first().attr('alt') ?? '';
-        const chapter = $('p.a-h.item-chapter > a.text-nowrap', manga).first().text().trim();
+    for (const item of $('.itemupdate.first', '.doreamon').toArray()) {
+        const title = $('h3', item).text();
+        const chapter = $('.sts_1', item).first().text();
+        const update = $('i', item).first().text();
 
-        const time_string = $('p.a-h.item-chapter > i', manga).first().text().trim();
-        const time = calculateTime(time_string);
-
+        const time = calculateTime(update);
         if (time > 60) {
             break;
         }
@@ -42,7 +41,7 @@ export default async function fetchMangaNelo(): Promise<Array<MangaPacket>> {
         const mangapacket: MangaPacket = {
             Name: title,
             Chapter: chapter,
-            Source: 'manganelo'
+            Source: 'mangakakalot'
         };
 
         mangaNeloUpdates.push(mangapacket);
