+++ server/src/index.ts
@@ -1,14 +1,13 @@
 import { promises as fs } from 'fs';
-import { MangaPacket } from 'types/sourceEntries';
-import { UserJson } from 'types/userJson';
+import { UserJson, MangaPacket } from './types';
 import { checkCache } from './cache';
 // import fetchMangaDex from './SourceUpdates/MangaDex';
 import fetchMangaFox from './SourceUpdates/MangaFox';
 import fetchMangaLife from './SourceUpdates/MangaLife';
 import fetchMangaKakalot from './SourceUpdates/MangaKakalot';
 import fetchMangaNelo from './SourceUpdates/MangaNelo';
-import sendIfttt from './NotificationHandler/ifttt';
-import sendSpontit from './NotificationHandler/spontit';
+import sendIfttt from './NotificationServices/ifttt';
+import sendSpontit from './NotificationServices/spontit';
 import { fetchUserJson, removeExtraChars } from './utils';
 
 if (require.main === module) {
