+++ server/src/types/index.ts
@@ -1,0 +1,3 @@
+export * from './services';
+export * from './sourceEntries';
+export * from './userJson';
