+++ server/src/types/userJson.ts
