+++ setup-wizard/src/types/index.ts
@@ -1,0 +1,2 @@
+export * from './manualSetup';
+export * from './paperbackBackup';
