+++ setup-wizard/src/types/manualSetup.ts
